fix: functionality removed to content.go
